v0.8.0

+++ pages/index.js
@@ -26,17 +26,38 @@ export default function Home() {
     }
     getTodoMainList()
   }, [])
+  
   //* HANDLES THE DELETE BUTTON FUNCTIONS OF THE MAIN TODO LIST ITEM
   const handleDelete = (index) =>{
     //* Delete Function
-    const todosMainList = JSON.parse(localStorage.getItem('todosMain'));
-    const filtered = todosMainList.filter((item,indexTodo) => indexTodo!=index)
-    localStorage.setItem('todosMain',JSON.stringify(filtered));
-    setTodosMain(filtered)
+    if(confirm(`Are you sure you want delete!`)){
+      const todosMainList = JSON.parse(localStorage.getItem('todosMain'));
+      const filtered = todosMainList.filter((item,indexTodo) => indexTodo!=index)
+      localStorage.setItem('todosMain',JSON.stringify(filtered));
+      setTodosMain(filtered)
+    }
   }
   //* HANDLES THE ARCHIVE BUTTON FUNCTIONS OF THE MAIN TODO LIST ITEM
   const handleArchive = (index) => {
-    console.log("Clicked",index)
+    const todosMainList = JSON.parse(localStorage.getItem('todosMain'));
+    //* Get the archive Item.
+    const archiveItem = todosMainList[index];
+    //* Add it to the archive List, If the archive exist just push else, create one and push.
+    if(JSON.parse(localStorage.getItem('archivedList'))){
+      const archivedList = JSON.parse(localStorage.getItem('archivedList'));
+      archivedList.push(archiveItem);
+      localStorage.setItem('archivedList',JSON.stringify(archivedList));
+    }
+    else{
+      localStorage.setItem('archivedList',JSON.stringify([]));
+      const archivedList = JSON.parse(localStorage.getItem('archivedList'));
+      archivedList.push(archiveItem);
+      localStorage.setItem('archivedList',JSON.stringify(archivedList));
+    }
+    //* Delete it from the main list
+    const filtered = todosMainList.filter((item,indexTodo) => indexTodo!=index)
+    localStorage.setItem('todosMain',JSON.stringify(filtered));
+    setTodosMain(filtered)
   }
 
   //* HANDLES THE SUMBIT of CREATE A TODO
@@ -65,11 +86,24 @@ export default function Home() {
     }
   }
 
+  //* Show a random Text when the whole thing is empty
+  const showRandomText = () => {
+    const arrayOfHappyText = [
+      `Looks like you're all done for now 😊`,
+      `Ohh it looks empty here 🤔, Make a todo`,
+      `Are you sure, you're all done 😅`,
+      `Good job, you have no more MonkDos 🚀`,
+      `Statistics is where you can see how you're doing 👨‍💻`,
+    ]
+    const randomValue = Math.floor(Math.random()*arrayOfHappyText.length);
+    return arrayOfHappyText[randomValue];
+  }
+
   return (
     <Layout>
       <div className="your-todos">
         <div className="header-text justpadding">
-          <p>Your MonkDos</p>
+          <p>{todosMain && (todosMain.length <= 1) ? `Your MonkDo` : `Your MonkDos`}</p>
         </div>
         <div className="createTodoInput">
           <input
@@ -92,8 +126,21 @@ export default function Home() {
           })} */}
           
           {todosMain ? todosMain.map((todoItem,index) => {
-            return <MainTodoItem key={index} index={index} handleDelete={handleDelete} handleArchive={handleArchive} title={todoItem.title} date={`${todoItem.day}-${todoItem.month}-${todoItem.year}`}/>
+            return <MainTodoItem key={index} index={index} handleDelete={handleDelete} handleArchive={handleArchive} title={todoItem.title} date={`[${
+              todoItem.hour > 12
+                ? `0${todoItem.hour.toString()}`
+                : `0${todoItem.hour.toString()}`
+            }:${
+              todoItem.minute < 10
+                ? `0${todoItem.minute.toString()}`
+                : todoItem.minute
+            } ${todoItem.hour > 12 ? `PM` : `AM`}] [${todoItem.day}-${todoItem.month}-${todoItem.year}]`}/>
           }) : `Loading`}
+          {todosMain && todosMain.length ===0 ? 
+          <div className="text-xl text-center px-10 md:px-5">
+            {showRandomText()}
+          </div>
+          : ``}
         </div>
 
       </div>
